sindy optimizer, added to sindy_autoencoder function
