--- aesindy/config_template.py
@@ -1,1 +1,0 @@
-ROOTPATH='todo'
