+++ requirements.txt
@@ -8,5 +8,4 @@ pysindy
 scikit_learn
 scipy
 tensorflow
-tensorflow_macos
 tqdm
