+++ testcases/pendulum_basic.py
@@ -11,21 +11,30 @@ os.environ["CUDA_VISIBLE_DEVICES"] = "0"
 
 params['model'] = 'pendulum'
 params['case'] = 'basic'
-params['system_coefficients'] = [9.8, 10]
+params['system_coefficients'] = [10, 10]
 params['noise'] = 0.0
 params['input_dim'] = 80
-params['dt'] = np.sqrt(params['system_coefficients'][0]/params['system_coefficients'][1])/params['input_dim']/10
-params['tend'] = 4
-params['n_ics'] = 30
+params['dt'] = np.sqrt(params['system_coefficients'][0]/params['system_coefficients'][1])/20
+params['tend'] = 30 
+params['n_ics'] = 35
 params['poly_order'] = 1
 params['include_sine'] = True
 params['fix_coefs'] = False
+params['coefficient_initialization'] = 'true'
 
-params['save_checkpoints'] = True 
+params['sindy_learning_rate'] = 0.1 
+params['learning_rate'] = 1e-3
+
+params['activation'] = 'elu'
+params['widths_ratios'] = [0.5, 0.25]
+params['use_bias'] = True 
+
+params['save_checkpoints'] = False
 params['save_freq'] = 1 
 
 params['print_progress'] = True
-params['print_frequency'] = 5 
+params['print_frequency'] = 1
+params['verbose'] = 0
 
 # training time cutoffs
 params['max_epochs'] = 3000
@@ -33,14 +42,13 @@ params['patience'] = 70
 
 # loss function weighting
 params['loss_weight_rec'] = 0.3
-params['loss_weight_sindy_z'] = 0.001 
+params['loss_weight_sindy_z'] = 0.0001 
 params['loss_weight_sindy_x'] = 0.001
 params['loss_weight_sindy_regularization'] = 1e-5
 params['loss_weight_integral'] = 0.1  
 params['loss_weight_x0'] = 0.01 
 params['loss_weight_layer_l2'] = 0.0
 params['loss_weight_layer_l1'] = 0.0 
-
 
 
 S = SynthData(model=params['model'],
